+++ lib/linux/ExploreException.h
@@ -83,3 +83,22 @@ public:
 	}
 
 };
+
+// class to throw exception when no BT service is available
+class ExploreNoBluetoothException : std::exception
+{
+private:
+	std::string message;
+public:
+	ExploreNoBluetoothException(std::string message) NOEXCEPT
+	{
+		this->message = message;
+
+	}
+
+	virtual const char* what() const NOEXCEPT
+	{
+		return message.c_str();
+	}
+
+};