+++ CHANGELOG.rst
@@ -2,6 +2,17 @@
 Changelog
 =========
 
+3.0.1 (28.08.2024)
+------------------
+* Improve device connectivity
+* Add text marker support
+
+3.0.0 (7.6.2024)
+------------------
+* Improve Mac OSX support
+* Add Support for Explore Pro
+* Robust data parsing
+
 2.0.0 (5.1.2024)
 ------------------
 * Add Mac OSX support